Implemented using Origin information. Cleaned up code.
